enable line numbers (for the upteenth time)
I've gone back to using line numbers again as I find myself guessing how
many lines up/down to delete. Whereas, with the line numbers I no longer
have to guess.

Jury is still out on whether I want relative line numbers, or just
absolute like I'm currently using.

I also wanted a way to show where the cursor is on screen easier when
I'm navigating between files. It can get lost since the colour of the
cursor is the same as most text. To fix this, I've added a line number
highlight for the line that the cursor is on. This is much more obvious
than having nothing, but much less intrusive than highlighting the
entire line the cursor is on.

Finally, make sure not to show line numbers in markdown files as it gets
in the way of simply writing text.