add better markdown support
Discovered a new plugin called "markdown" that allows for the rendering
of markdown files within the neovim window.

This is what I've been looking for, and it appears to conform to the
conceallevel option that I've set in the ftplugin folder for markdown
files.

Will continue to configure this for the future, but immediate feelings
is that it's pretty good.